requirements.txt
termcolor
python-dotenv
